Remove NIX_PATH check for ipp-crypto build
Instead just check if the submodule is empty, and if it is then proceed
as it was done before.

A better approach could be to not only verify if the ipp-crypto
submodule is not empty but to check that it is up-to-date, and if it is
either empty or out-of-date then proceed as before, meaning, init the
submodule, or clone.

An even better approach: why not have ipp-crypto as an external
dependency like any other dependency? Perhaps there are good reasons not
to do so.